Adding properties, abstracting out the Server.
Added a Server.java class, to represent the remote server to
be tested, and added the ability to load its identity through a
properties mechanism.  Copy "skeleton-properties" to a new file
(such as "your-host-name.properties") and then modify its contents
to indicate the location of the test server.  Later updates will let
you specify the properties file on the command line.