Updated the Request schema to have correct Links.
Some of the fields ("created_by", etc.) for the Request.js schema
were listed as having type "User" or other top-level schemas, instead
of Link types to the corresponding resources.  This may change in
the future, but for now the schema file has been updated to reflect
the Broker's current practice.